Fixed a bunch of issues with Process. Overriding fs Added keepAlive() and letDie() methods to Process All tests passing

+++ NodeDroid-library/jni/JSFunction.hpp
@@ -139,7 +139,7 @@ class JSFunction : public JSValue<T> {
 
                 argsArr = env->NewLongArray(argumentCount);
                 args = new jlong[argumentCount];
-                for (size_t i=0; i<argumentCount; i++) {
+                for (int i=0; i<argumentCount; i++) {
                     args[i] = reinterpret_cast<long>(JSValue<Value>::New(
                         JSValue<T>::m_context, info[i]));
                 }
